requirements.txt
# AMLAC Robot - Python Dependencies
# Install all packages with: pip3 install -r requirements.txt

# TensorFlow Lite Runtime for ML inference
tflite-runtime>=2.14.0

# Camera support
picamera2>=0.3.12

# GPIO control
RPi.GPIO>=0.7.1

# Color sensor (TCS34725)
adafruit-circuitpython-tcs34725>=3.3.0

# I2C communication
smbus2>=0.4.2

# LCD display
RPLCD>=1.3.1

# Load cell (HX711)
hx711>=1.1.0

# GPS parsing
pynmea2>=1.19.0
pyserial>=3.5

# Image processing
numpy>=1.24.0
Pillow>=10.0.0

# Additional dependencies
adafruit-blinka>=8.20.0
adafruit-circuitpython-busdevice>=5.2.0

